Replaced NewtonSoft package with System.Text.Json
This removes the `Newtonsoft.Json 12.0.3` dependency that causes issues when using other packages that reference the newer `13.0.1` version of Newtonsoft (see #443 for the same issue with the older Newtonsoft 11.0/12.0 conflict).

I've replaced `Newtonsoft.Json` completetly with `System.Text.Json 5.0.0`.  
Unfortunately this results in a bit of ugly code in the `Dotnet.Script.Core.Scaffolder` class to update a JSON property, as this version of `System.Text.Json` doesn't allow manipulation of the JSON DOM - I've had to manually re-write the  JSON.

The impending `6.0.0` version will resolve that as it has writable DOM support, making it possible for this code to be much simpler (much like the Newtonsoft version) - I'll update this again once that version is out of beta.