Added '--cache-path' command line option
Added a new command-line option "--cache-path" (or "-p") to allow users to specify the temporary directory used to build the script. This is equivalent to the 'DOTNET_SCRIPT_CACHE_LOCATION' environment but is far more convenient when scripts are run as part of MSBuild.  In particular, parallel runs of a script can specify different cache paths preventing the "script.csproj is in use" failure that can happen. (Re: issues #507, #596, #540)